Add files via upload
This the pdf file of my Dissertation for my MSc in Data Science. It is an in-depth time series analysis of the Chinese economy, encompassing key indicators such as trade, GDP, FDIs, and inflation. I employed statistical models, including the use of the statsmodels library and ARIMA to forecast prices and identify significant trends, demonstrating a comprehensive understanding of economic dynamics and proficiency in advanced analytical techniques. It was developed during the fall of 2023 and was maked with a distinction.